dont fail for files > 2GB
git-svn-id: https://svn.xiph.org/experimental/giles/rogg@16007 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ rogg.c
@@ -110,7 +110,7 @@ void rogg_page_get_length(unsigned char *p, int *length)
 }
 
 /* scan for the capture pattern */
-unsigned char *rogg_scan(unsigned char *p, int len)
+unsigned char *rogg_scan(unsigned char *p, long len)
 {
   unsigned char *end = p + len - 4;
